新增敏感词分类接口

+++ src/main/java/com/kinghy/rag/controller/UserController.java
@@ -136,7 +136,7 @@ public class UserController {
      */
     @PostMapping("/status/{status}")
     @Operation(summary = "status",description = "启用禁用账号")
-    public BaseResponse startOrStop(@PathVariable Integer status,Long id){
+    public BaseResponse startOrStop(@PathVariable Integer status,Integer id){
         log.info("启用禁用员工账号：{},{}",status,id);
         userService.startOrStop(status,id);
         return ResultUtils.success("禁用成功");
